requirements.txt
# FastAPI and Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
websockets>=12.0
python-multipart>=0.0.6
jinja2>=3.1.0

# OpenAI Agents SDK and AI
openai-agents>=0.1.0
openai>=1.87.0
pydantic>=2.5.0

# GCP Integration
google-cloud-logging>=3.8.0
google-auth>=2.23.4
google-cloud-core>=2.3.3

# Email Service
sendgrid>=6.10.0

# Data Processing and Utils  
python-dateutil>=2.8.2
numpy>=1.24.3
pandas>=2.0.3

# Machine Learning for Intelligent Anomaly Detection
scikit-learn>=1.3.0
scipy>=1.11.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.3

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Logging and Monitoring
structlog>=23.2.0
rich>=13.7.0
